Rename combo package to integer
