+++ src/config/Settings.java
@@ -82,7 +82,7 @@ public class Settings {
 	public static boolean cacheDerivatives = true;
 
 	/**
-	 * Forces functions in {@link functions.unitary.combo} to return integers when using approximations
+	 * Forces functions in {@link functions.unitary.integer} to return integers when using approximations
 	 */
 	public static boolean enforceIntegerOperations = true;
 
@@ -112,7 +112,7 @@ public class Settings {
 	public static SolverType defaultSolverType = SolverType.NEWTON;
 
 	/**
-	 * Denotes the default implementation of {@link functions.unitary.combo.Factorial} to be used
+	 * Denotes the default implementation of {@link functions.unitary.integer.Factorial} to be used
 	 */
 	public static FactorialType defaultFactorial = FactorialType.RECURSIVE;
