+++ src/config/cas.properties
@@ -35,7 +35,7 @@ distributeExponents = true
 # Denotes whether or not the derivatives of functions should be cached when created
 cacheDerivatives = true
 
-# Forces functions in {@link functions.unitary.combo} to return integers when using approximations
+# Forces functions in {@link functions.unitary.integer} to return integers when using approximations
 enforceIntegerOperations = true
 
 # Denotes whether methods in {@link tools.singlevariable} should exit if the result is within a certain proximity of the target. Improves performance at the cost of accuracy.
