+++ src/functions/unitary/integer/RFactorial.java
@@ -1,4 +1,4 @@
-package functions.unitary.combo;
+package functions.unitary.integer;
 
 import config.Settings;
 import functions.GeneralFunction;
